+++ MK4duo/src/core/commands/gcode/gcode.h
@@ -61,6 +61,7 @@
 #include "config/m305.h"                  // Set thermistor and ADC parameters
 #include "config/m306.h"                  // Set Heaters
 #include "config/m595.h"                  // Set AD595 offset & Gain
+#include "config/m569.h"                  // Set Stepper Direction
 #include "config/m900.h"                  // Set and/or Get advance K factor
 #include "config/m906.h"                  // Set Alligator motor currents or Set motor current in milliamps with have a TMC2130 driver
 #include "config/m907.h"                  // Set digital trimpot motor current